+++ examples/neptune_tracking/steps/evaluator/evaluator_step.py
@@ -2,7 +2,7 @@ import numpy as np
 import tensorflow as tf
 
 from zenml.steps import step
-from zenml.integrations.neptune.neptune_utils import get_neptune_run, neptune_step
+from zenml.integrations.neptune.experiment_trackers.run_state import get_neptune_run, neptune_step
 
 
 @neptune_step
